requirements.txt
python-dotenv 
urllib3
Flask
Werkzeug
MarkupSafe
mysql-connector
requests
gunicorn
flask-cors